moved the footer containing add criteria outside the scroll wrapper
